requirements.txt
Flask-Bootstrap==3.3.7.1
Flask==3.1.2
Flask-Reuploaded==1.4.0
Flask-WTF==1.2.2
label-studio==1.21.0 
matplotlib==3.10.5
nltk==3.9.1
numpy==1.26.4
opencv-python==4.11.0.86
pillow==12.0.0
pyarrow==18.1.0
python-dotenv==1.2.1
python-slugify==8.0.4
PyYAML==6.0.2
regex==2024.11.6
torch==2.8.0
tqdm==4.67.1
ultralytics==8.3.186
urllib3==2.5.0
WTForms==3.2.1